Switch to more modern Gradle plugins
This switches to using the latest version of Gradle, updates the
project build tools, and puts all version information in a central
location rather than repeating it in manifest files. This corrects
a build error introduced in a recent version of Android Studio.

Affects: https://jira.nypl.org/browse/SIMPLY-2224